LIBFCREPO-1100. Force version 4.5.13 of the Apache HTTP Client.
Because the vocab.lib.umd.edu server uses a 308 redirect for http-to-https, must use a client version that recognizes that code. Apache HTTP Client 4.5.12 is the first version to do so, and 4.5.13 is the current 4.5.x version as of this commit.

https://issues.umd.edu/browse/LIBFCREPO-1100
